updated all required files

+++ src/addVideo.php
@@ -1,4 +1,5 @@
 <?php
+    
     error_reporting(E_ALL);
     ini_set('display_errors', 'On');
     include 'inventory.php';
@@ -6,17 +7,40 @@
     $name = $_POST['name'];
     $category = $_POST['category'];
     $length = $_POST['length'];
+    $unique = "SELECT name FROM inventory";
+    $filterUnique = mysqli_query($mysqli, $unique);
     
-    /* Error handling */
+    //Error checking
+    
+    //Movie name must be unique
+    while($filterName = $filterUnique->fetch_assoc()){
+        foreach($filterUnique as $filterName){
+            if($filterName['name'] == $_POST['name']){
+                echo "This movie already exists";
+                die();
+            }
+        }
+    }
+
+    //Movie name is required
     if(strlen($name) == 0){
         echo 'A movie name is required.';
         die();
     }
     
+    //Movie length must be numeric
     if(!is_numeric($length)){
         echo 'Length must be numeric.';
         die();
     }
+    
+    //Movie length must be positive
+    if($length < '0'){
+        echo 'Length must be positive.';
+        die();
+    }
+    
+    //http://us2.php.net/manual/en/mysqli.quickstart.prepared-statements.php
     
     /* Non-prepared statement
     if (!$mysqli->query("DROP TABLE IF EXISTS inventory") || !$mysqli->query("CREATE TABLE inventory(id INT PRIMARY KEY AUTO_INCREMENT, name VARCHAR(255) UNIQUE NOT NULL, category VARCHAR(255), length INT UNSIGNED, rented BOOL DEFAULT NULL)")) {
@@ -38,5 +62,8 @@
         echo "Execute failed: (" . $stmt->errno . ") " . $stmt->error;
     }
     
+    echo "<a href = 'http://web.engr.oregonstate.edu/~santokis/cs290-assignment4-part2/inventory.php'>refresh page</a>";
+    
     $stmt->close();
+    
 ?>
